Add diagnostic script for DEM730P RS485 communication and enhance scanning capabilities
- Introduced `debug_powermeter.py` for low-level diagnostics of RS485 communication with the DEM730P power meter, displaying raw bytes sent and received.
- Updated `scan_powermeter.py` to support a wider range of Modbus addresses (1-254) and added a fast mode for quicker scanning.
- Modified the ingestion worker to change the conflict resolution strategy for station inserts to "DO NOTHING".
- Added a new consumer script for Spark Structured Streaming to read sensor events from Kafka, compute windowed statistical features, and write results to PostgreSQL.
- Created a Docker Compose setup for Kafka and Kafka UI to facilitate local development and testing.
- Implemented MLflow setup script to initialize experiments and model registrations for anomaly detection and drift adaptation phases.
- Added producer script to read sensor measurements from PostgreSQL and publish them to Kafka, supporting both replay and live modes.
- Updated requirements for Phase 1 streaming foundation, ensuring compatibility with necessary libraries.
- Defined schema for `windowed_features` table to store aggregated sensor data.
- Adjusted PostgreSQL schema for measurements to change the data type of energy and flow metrics for consistency.

+++ schema_timescaledb.sql
@@ -45,7 +45,7 @@ CREATE TABLE IF NOT EXISTS measurements (
     -- Power metrics
     voltage DOUBLE PRECISION,
     power DOUBLE PRECISION,
-    energy BIGINT,
+    energy DOUBLE PRECISION,
 
     -- Flow metrics
     flow_lmin DOUBLE PRECISION,
